+++ src/OAuth/Introspection/OAuthIntrospectionException.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace BWH\Auth\OAuth\Introspection;
+
+use RuntimeException;
+
+final class OAuthIntrospectionException extends RuntimeException
+{
+}